+++ som_ov_production_data/som_ov_production_data.py
@@ -8,6 +8,10 @@ from som_ov_users.decorators import www_entry_point
 from som_ov_users.exceptions import (
     NoSuchUser,
 )
+from som_ov_installations.exceptions import (
+    UnauthorizedAccess,
+    ContractNotExists,
+)
 
 
 class SomOvProductionData(osv.osv_memory):
@@ -29,6 +33,7 @@ class SomOvProductionData(osv.osv_memory):
             context = {}
 
         contracts = self._get_user_contracts(cursor, uid, username, context)
+
         measures = {'data': []}
         for contract in contracts:
             contract_data = self._get_production_measures(cursor, contract, first_timestamp_utc, last_timestamp_utc)[0][0]
@@ -36,6 +41,47 @@ class SomOvProductionData(osv.osv_memory):
             measures['data'].append(contract_data)
 
         return measures
+
+    @www_entry_point((
+        NoSuchUser,
+        UnauthorizedAccess,
+        ContractNotExists,
+    ))
+    def measures_single_installation(
+        self, cursor, uid,
+        username,
+        contract_number,
+        first_timestamp_utc,
+        last_timestamp_utc,
+        context=None
+    ):
+
+        if context is None:
+            context = {}
+
+        contract = self._get_user_contract(cursor, uid, username, contract_number)
+
+        measures = {'data': []}
+        contract_data = self._get_production_measures(cursor, contract, first_timestamp_utc, last_timestamp_utc)[0][0]
+        contract_data['foreseen_kwh'] = self._get_forecast_measures(cursor, uid, contract.cil.id, first_timestamp_utc, last_timestamp_utc, context)
+        measures['data'].append(contract_data)
+
+        return measures
+
+    def _get_user_contract(self, cursor, uid, username, contract_number):
+        polissa_obj = self.pool.get('giscere.polissa')
+        contract_id = polissa_obj.search(cursor, uid, [('name', '=', contract_number)])
+        if not contract_id:
+            raise ContractNotExists()
+        contract = polissa_obj.browse(cursor, uid, contract_id[0])
+
+        if contract.titular_nif != username:
+            raise UnauthorizedAccess(
+                username=username,
+                resource_type='Contract',
+                resource_name=contract_number,
+            )
+        return contract
 
     def _get_user_contracts(self, cursor, uid, username, context):
         installation_obj = self.pool.get('som.ov.installations')